Move automount package out of pkg/provision/workspace
