+++ pkg/provision/automount/common.go
