+++ pkg/provision/automount/git-user-credentials.go
