+++ pkg/provision/automount/git-user-credentials_test.go
