+++ pkg/provision/automount/secret.go
